make pay-to-open work with non-trampoline AMP
We now accept incoming AMP payments that are paid by a combination of
htlcs and pay-to-open requests. All pay-to-open requests are combined in
a single request.

More symmetry would have been nice between legacy/trampoline (e.g. for
the `SendPayToOpen.origin` field) but it didn't seem possible.

Also added an expiry timestamp to prevent race conditions at the UI
level.